jsplus ce que j'ai fait

+++ libft_rendu_pdf2022/ft_libc/ft_memcpy.c
@@ -2,10 +2,10 @@
 
 void    *ft_memcpy(void *dest, const void *src, size_t n)
 {
-    int    i;
+    size_t   i;
 
     i = -1;
-    while (i++ < (int)n)
+    while (i++ < n)
         ((unsigned char *)dest)[i] = ((unsigned char *)src)[i];
     return (dest);
 }
